+++ platform/nxp/gen_1/i2c_slave.c
@@ -211,7 +211,7 @@ static enum result i2cGet(void *object, enum ifOption option, void *data)
   switch (option)
   {
     case IF_ADDRESS:
-       *(uint8_t *)data = ADR_ADDRESS_VALUE(reg->ADR0);
+       *(uint16_t *)data = ADR_ADDRESS_VALUE(reg->ADR0);
       return E_OK;
 
     case IF_POSITION:
@@ -234,9 +234,9 @@ static enum result i2cSet(void *object, enum ifOption option, const void *data)
   switch (option)
   {
     case IF_ADDRESS:
-      if (*(const uint8_t *)data <= 127)
+      if (*(const uint16_t *)data <= 127)
       {
-        reg->ADR0 = ADR_ADDRESS(*(const uint8_t *)data);
+        reg->ADR0 = ADR_ADDRESS(*(const uint16_t *)data);
         return E_OK;
       }
       else
